Split AES helper functions into its own file. Considering adding a symmetric crypto a package. Added comments to cryptospecials/rsavrf.go to indicate that it uses SHA2 not SHA1 when calulating MGF1. This will matter b/c SHA2 is non-standard when using MGF1
